Refactor: extract implementations from schema.service.ts
Extracted implementations to schema.service.impl.ts
Did not extract ReferencePorpertyImpl and ContainmentPropertyImpl but
exported them to avoid circular dependency with schema.service.impl.ts
